### 2018-12-07 #1121 @whedon
[PRE REVIEW]: UWGeodynamics: A teaching an research tool for numerical geodynamic modelling
**Submitting author:** @rbeucher (<a href="http://orcid.org/0000-0003-3891-5444">Romain Beucher</a>)
**Repository:** <a href="https://github.com/underworldcode/UWGeodynamics" target ="_blank">https://github.com/underworldcode/UWGeodynamics</a>
**Version:** v1.0.0
