### 2018-12-07 #1121 @lheagy
Hi @arfon, I am happy to be an editor on this!

It looks like the paper is in a directory called `joss` - can this be communicated to whedon? or does the paper need to be in root or a `paper` directory
