### 2018-12-08 61780af
remove README file, change link to rst format
