### 2018-12-10 #1121 @lheagy
👋 @arfon: can the paper be in a folder called "joss" or does it need to be in root or "paper"
