### 2018-12-10 06ddcfa
Fix stressBCs
