### 2018-12-10 d778fe2
Get long description from README.rst
