+++ setup.py
@@ -15,7 +15,7 @@ VERSION             = '%d.%d.%d' % (MAJOR, MINOR, MICRO)
 here = path.abspath(path.dirname(__file__))
 
 # Get the long description from the README file
-with open(path.join(here, 'README'), encoding='utf-8') as f:
+with open(path.join(here, 'README.rst'), encoding='utf-8') as f:
     long_description = f.read()
 
 
