### 2018-12-10 #1121 @arfon
@lheagy the location is fine. The only issue I can see is L19: `orcid:0000-0003-3891-5444` <- this needs a space between `orcid:` and the value.
