### 2018-12-12 faf490e
Fix issues with temperature field when using the material condition and when running the code in parallel
