### 2018-12-13 f93fbfc
Fix shape as nodeSet, was causing issue in parallel, CanonBall example was broken
