### 2018-12-13 b4fe897
Explicitely define gravity vector
