+++ joss/paper.md
@@ -16,7 +16,7 @@ tags:
   - stokes
 authors:
  - name: Romain Beucher
-   orcid:0000-0003-3891-5444
+   orcid: 0000-0003-3891-5444
    affiliation: "1"
  - name: Louis Moresi
    orcid: 0000-0003-3685-174X
