+++ joss/paper.md
@@ -52,7 +52,7 @@ authors:
    orcid: 0000-0002-3484-7985
    affiliation: "5"
  - name: Sara Morón
-   orcid: 0000-0002-1270-437
+   orcid: 0000-0002-1270-4377
    affiliation: "1"
 
 affiliations:
