### 2018-12-14 6d0b9d9
Refactor some internal function so that the solver options are easier to tune. This removes the get_stokes_solver method The solver object is initialized during the init_model step The SLE is rebuilt at each time step. Solver options can be changed
