### 2018-12-14 4dac9db
This is also necessary for the restart in parallel
