### 2018-12-14 ddbd20c
The viscosity limiter only worked when both min and max viscosities were defined, this allows for both or only one to be defined
