+++ UWGeodynamics/_model.py
@@ -2560,11 +2560,11 @@ class _RestartFunction(object):
         if uw.rank() == 0:
             swarm_file = os.path.join(self.restartDir, "swarm-%s.h5" % step)
             with h5py.File(swarm_file, "r") as h5f:
-                Model.time = u.Quantity(h5f.attrs.get("time"))
+                Model._ndtime = nd(u.Quantity(h5f.attrs.get("time")))
         else:
-            Model.time = None
+            Model._ndtime = None
 
-        MPI.COMM_WORLD.bcast(Model.time, root=0)
+        Model._ndtime = MPI.COMM_WORLD.bcast(Model._ndtime, root=0)
 
         if uw.rank() == 0:
             print(80 * "=" + "\n")
