### 2018-12-15 fbe5a57
Fix bcast of Model.time
