### 2018-12-15 457bb2c
Do not rebuild the solver unless specified
