### 2018-12-15 7f23ac4
Point travis badge to master
