### 2018-12-16 835583a
Fix, failed when there is no passive tracers to reload
