### 2018-12-16 f59ae51
Fix, failed when there is no passive tracers to reload
