### 2018-12-16 4bcccc7
Move tractionField to the initialize method. (For restarting model with stressBCs)
