### 2018-12-17 96c3560
Add timeField to the restart variables
