### 2018-12-17 8543fc5
Do not get a solver at the end of Model.init, the rationale behind this was to have a solver class build to look at the options available...
