### 2018-12-18 #1136 @whedon
[REVIEW]: UWGeodynamics: A teaching an research tool for numerical geodynamic modelling
**Submitting author:** @rbeucher (<a href="http://orcid.org/0000-0003-3891-5444">Romain Beucher</a>)
**Repository:** <a href="https://github.com/underworldcode/UWGeodynamics" target ="_blank">https://github.com/underworldcode/UWGeodynamics</a>
**Version:** 2.7.7

Reviewers, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
