### 2018-12-18 #1136 @lheagy
Many thanks @techas, @tth030 for being willing to review! 🎉

First of all, if you have not signed up to be a reviewer with JOSS before, please add yourself here: http://joss.theoj.org/reviewer-signup.html/. This will allow you to check-off the boxes above.

There is a checklist for each of you to help guide the review. If there are items that you see are missing or could be improved, please either comment here or create an issue in the [target repository](https://github.com/underworldcode/UWGeodynamics) and reference this issue by including `openjournals/joss-reviews#1136` in the text of the issue.

Please let me know if you have any questions or if I can clarify anything.
