### 2018-12-20 dfadc01
Activate examples tests
