### 2018-12-20 c87b8c5
deactivate test in 3D for now, too hard on 1 CPU even for small models
