### 2018-12-30 9648c32
Call print_petsc_options() from all procs
