### 2019-01-03 eda3bd4
comment out Petsc option printing
