### 2019-01-07 #1136 @lheagy
👋 Hi @techas, @tth030, I hope you had a wonderful time over the holidays! Now that the new year is starting to get going, I wanted to follow up and ask when you think you will have time to complete the review? Ideally, we would appreciate if you could do so in the next 2 weeks. Please let me know if you have any questions!
