### 2019-01-14 bb71217
Add option to only advect passive tracers vertically
