### 2019-01-14 7bcc582
SLE cannot be dynamically assigned to the solver object. This caused an issue when restarting models. Force the build of a new solver object when doing a restart
