### 2019-01-14 fd6a8ee
Revert "Add option to only advect passive tracers vertically"
This reverts commit bb712178e3d4b7f8c6cb3697103a671fec9f05da.
