### 2019-01-15 7f1864f
Fix grid spacing in linkage, nx, ny were the number of cells instead of nodes
