### 2019-01-15 64521fb
Add option to only move passive tracers in z
