### 2019-01-15 6b22bf1
petsc_options is fairly recent, pass if the Underworld version does not have it
