### 2019-01-17 dc64c66
print_petsc_option only works with 2.7 and above
