### 2019-01-20 0b34fca
Pass zOnly when restarting passive tracers, particles get advected otherwise
