### 2019-01-20 73b0a3c
Add meltField to default outputs
