### 2019-01-24 634dc59
Initialize tracked swarm variable to 0.
