### 2019-01-29 28cdd79
Fix restart badalnds
