### 2019-01-29 2e8df23
Fix restart badalnds
