### 2019-02-03 #1136 @tth030
Installation: Does installation proceed as outlined in the documentation?

** NO

After many tests, I was unable to test UWGeodynamics.

LOCAL INSTALLATION
I worked under Mac Os x 10.12.6.
Underworld2 is required to run UWGeodynamics. I believe that my system is well set. I use Spack to deal with my environment.  I went through the installation using python2.7.15 but ipython5.8.0 is freezing when importing underworld2...
Using Python3.6.5 I got this message ""PETSc was configured with one OpenMPI mpi.h version but now appears to be compiling using a different OpenMPI mpi.h version"" when running "scons.py" during underworld2 installation.
Unfortunately, I do not have so much time to go further other tests for a local installation.

DOCKER INSTALLATION

To be honest this was painful to have to install Docker (>> 1.8 Gb !!!!!!)...
Docker then ask for privileges on my laptop...
We then have a huge black box installation in docker >> 1-2 Gb ...
This is a crazy way only to install python packages...
--port 8888:8888 \ is not a recognized option by docker...

The installation process has to be improved. I understand that authors want simplify users life thanks to the use of a container because of underworld installation. It looks like underworld2 to be really sensitive to installation process (compiler, versions...?) and UWGeodynamics can only be run if underworld2 is installed.

I stop my review here. I apologize to JOSS and authors but I do not have enough time to go further this review.

Regards,
