### 2019-02-04 #1136 @lheagy
Thanks @tth030 for taking time to work through this. I can find someone else to step in as we move forward.

@rbeucher, I appreciate that some of the software installation steps might depend on how Underworld2 distributes software. One suggestion for simplifying things and demonstrating that the software can be installed would be to use binder: https://mybinder.org/ (which can take docker files) and provide an example notebook that uses UWGeodynamics (this could be in the same repository or a different one - however you prefer to organize things). The reviewer can then work through the installation steps you provide or use the example hosted on binder to complete the review. The docs on binder are here (https://mybinder.readthedocs.io/en/latest/). I am happy to answer any questions about getting up and running on binder.
