### 2019-02-04 #1136 @rbeucher
Thanks for your time @tth030. I appreciate that the Underworld dependencies, mainly petsc and hdf can be hard to compile, especially if one is not used to do that sort of things. It indeed requires skills and I am sorry you could not get it to work. This is why we provide  docker image. The Docker team provides detailed instructions on how to install docker. Now the image is indeed quite big. This is because it contains an OS and the full stack of dependencies required to run UWGeodynamics models, not just a python package. I wish you good luck in sorting out your busy schedule.
