### 2019-02-05 #1136 @rbeucher
Hi @lheagy,

I have set up a binder https://mybinder.org/v2/gh/rbeucher/UWGeodynamics-binder/master
I will add the link to the main repository.
I can suggest 2 reviewers:
Anthony Jourdon, University of Toulouse France, anthony.jourdon@get.omp.eu
Florian Wellman, Aachen University, wellmann@aices.rwth-aachen.de

Regards,
Romain Beucher
