### 2019-02-05 588c100
Add pandas to the Dockerfile
