### 2019-02-06 d169b79
Do not round the projMaterialField
