+++ UWGeodynamics/_model.py
@@ -396,9 +396,6 @@ class Model(Material):
     def projMaterialField(self):
         """ Material field projected on the mesh """
         self._materialFieldProjector.solve()
-        self._projMaterialField.data[:] = np.rint(
-            self._projMaterialField.data[:]
-        )
         return self._projMaterialField
 
     @property
