### 2019-02-06 b4782ae
Coordinates is a vector not a scalar, fix this
