### 2019-02-06 08af590
Deprecate rcParams solver and penalty
