### 2019-02-06 cb825c9
Add projStressField as default outputs
