### 2019-02-07 0d12c6b
Change version to 2.6.1, to match UW version number
