+++ docs/source/conf.py
@@ -26,7 +26,7 @@ author = 'Romain Beucher'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '0.9.8'
+release = '2.7'
 
 
 # -- General configuration ---------------------------------------------------
