### 2019-02-18 26cc11c
Some change to handles nodeSets as kinematic conditions (important for indenters), also make sure that each index is unique in the indexSets used to define boundary conditions
