### 2019-02-25 32b2640
Fix issue when passing uw.fn to 2nd and 3rd dimension
