### 2019-02-25 9f0010c
Fix issues when using an UW function on 2nd and 3rd dimensions
