### 2019-02-27 #1136 @lheagy
Many thanks @flohorovicic for being willing to review! If you haven't done so already, please accept the invitation from: https://github.com/openjournals/joss-reviews/invitations. This will allow you to check off the tick-boxes in the main thread above. If there are items that you see are missing or could be improved, please either comment here or create an issue in the [target repository](https://github.com/underworldcode/UWGeodynamics) and reference this issue by including `openjournals/joss-reviews#1136` in the text of the issue.

The process for installing some of the dependencies is a bit involved, but @rbeucher has set up a binder where you can test-drive the software: https://mybinder.org/v2/gh/rbeucher/UWGeodynamics-binder/master

Please don't hesitate to reach out if you have any questions. Thanks!
