+++ UWGeodynamics/_utils.py
@@ -13,6 +13,7 @@ from .Underworld_extended import Swarm
 from scipy import spatial
 from mpi4py import MPI
 
+
 class PhaseChange(object):
 
     def __init__(self, condition, result):
@@ -81,7 +82,7 @@ class PassiveTracers(object):
             velocityField=self.velocityField, order=2)
 
         indices = np.arange(self.swarm.particleLocalCount)
-        rank = uw.rank()
+        rank = uw.mpi.rank
         ranks = np.repeat(rank, self.swarm.particleLocalCount)
         pairs = np.array(list(zip(ranks, indices)), dtype=[("a", np.int32),
                                                            ("b", np.int32)])
@@ -185,7 +186,7 @@ class PassiveTracers(object):
 
         sH = self.swarm.save(swarm_fpath, units=u.kilometers, time=time)
 
-        if uw.rank() == 0:
+        if uw.mpi.rank == 0:
             filename = self.name + '-%s.xdmf' % checkpointID
             filename = os.path.join(outputDir, filename)
 
@@ -193,16 +194,16 @@ class PassiveTracers(object):
             string = uw.utils._xdmfheader()
             string += uw.utils._swarmspacetimeschema(sH, swarm_fname, time)
 
-        uw.barrier()
+        uw.mpi.barrier()
 
         # Save global index
         file_prefix = os.path.join(
             outputDir, self.name + '_global_index-%s' % checkpointID)
         handle = self.global_index.save('%s.h5' % file_prefix)
 
-        if uw.rank() == 0:
+        if uw.mpi.rank == 0:
             string += uw.utils._swarmvarschema(handle, "global_index")
-        uw.barrier()
+        uw.mpi.barrier()
 
         # Save each tracked field
         for field in self.tracked_field:
@@ -216,15 +217,15 @@ class PassiveTracers(object):
                 obj.data[...] = field["value"].evaluate(self.swarm)
             handle = obj.save('%s.h5' % file_prefix, units=field["units"])
 
-            if uw.rank() == 0:
+            if uw.mpi.rank == 0:
                 # Add attribute to xdmf file
                 string += uw.utils._swarmvarschema(handle, field["name"])
 
-        uw.barrier()
+        uw.mpi.barrier()
 
         # get swarm parameters - serially read from hdf5 file to get size
 
-        if uw.rank() == 0:
+        if uw.mpi.rank == 0:
             with h5py.File(name=swarm_fpath, mode="r") as h5f:
                 dset = h5f.get('data')
                 if dset is None:
@@ -244,7 +245,7 @@ class PassiveTracers(object):
             xdmfFH = open(filename, "w")
             xdmfFH.write(string)
             xdmfFH.close()
-        uw.barrier()
+        uw.mpi.barrier()
 
 
 class Balanced_InflowOutflow(object):
