+++ UWGeodynamics/_net/__init__.py
@@ -93,8 +93,8 @@ def PostGAEvent( category, action, label=None, value=None ):
         "ea" : action,          # Event Action. Required.
         "el" : label,           # Event label.
         "ev" : value,           # Event value.
-        "cm2": uw.nProcs(),     # Number of processes used. Stored into custom metric 2.
-        "cd5": str(uw.nProcs()),# Number of processes used. Stored into custom dim 2. Not sure if necessary.
+        "cm2": uw.mpi.nProcs(),     # Number of processes used. Stored into custom metric 2.
+        "cd5": str(uw.mpi.nProcs()),# Number of processes used. Stored into custom dim 2. Not sure if necessary.
         }
         import os
         # add user id if set
