### 2019-03-06 c1cdc15
Remove linkage. The Badlands coupling is now done through the Badlands class. This simplifies the code a lot
