+++ UWGeodynamics/linkage/linkage.py
@@ -1,343 +1,0 @@
-from __future__ import print_function,  absolute_import
-import sys
-from pyBadlands.model import Model as BadlandsModel
-from scipy.interpolate import griddata, interp1d
-from scipy.ndimage.filters import gaussian_filter
-import numpy as np
-from UWGeodynamics.scaling import COEFFICIENTS as scaling_coefficients
-from UWGeodynamics.scaling import Dimensionalize, UnitRegistry
-
-from tempfile import gettempdir
-from .uw_utilities import get_UW_velocities
-
-_tempdir = gettempdir()
-
-from mpi4py import MPI
-
-comm = MPI.COMM_WORLD
-size = comm.Get_size()
-rank = comm.rank
-
-class SPM(object):
-
-    def __init__(self, mesh, velocityField, swarm, materialField, airIndex,
-                 sedimentIndex, XML, resolution, checkpoint_interval,
-                 surfElevation=0., verbose=True, restartFolder=None,
-                 restartStep=None):
-
-        self.SECONDS_PER_YEAR = 31556925.9747  # Tropical year in seconds
-        self.verbose = verbose
-        self.restartStep = restartStep
-        self.restartFolder = restartFolder
-
-        # AutoScaling
-        self.scaleDIM = 1.0 / scaling_coefficients["[length]"].magnitude
-        self.scaleTIME = 1.0 / scaling_coefficients["[time]"].magnitude
-
-        self.mesh = mesh
-        self.velocityField = velocityField
-        self.swarm = swarm
-        self.material_index = materialField
-        self.airIndex = airIndex
-        self.sedimentIndex = sedimentIndex
-        self.resolution = resolution
-        self.surfElevation = surfElevation
-        self.checkpoint_interval = checkpoint_interval / self.scaleTIME / self.SECONDS_PER_YEAR
-
-        self.XML = XML
-
-        if rank == 0:
-            self.badlands_model = BadlandsModel()
-            self.badlands_model.load_xml(self.XML)
-            if self.restartStep:
-                self.badlands_model.input.restart = True
-                self.badlands_model.input.rstep = self.restartStep
-                self.badlands_model.input.rfolder = self.restartFolder
-                self.badlands_model.input.outDir = self.restartFolder
-                self.badlands_model.outputStep = self.restartStep
-
-        self.minCoord = self.mesh.minCoord
-        self.maxCoord = self.mesh.maxCoord
-
-        self.time_years = 0.
-        if self.restartStep:
-            # Parse xmf for the last timestep time
-            import xml.etree.ElementTree as etree
-            xmf = self.restartFolder+"/xmf/tin.time"+str(self.restartStep)+".xmf"
-            tree = etree.parse(xmf)
-            root = tree.getroot()
-            self.time_years = float(root[0][0][0].attrib["Value"])
-
-        self._tmp = _tempdir
-        self._demfile = self._tmp+"/dem.csv"
-
-        # Create Initial Flat DEM
-        if rank == 0:
-            self.dem = self._generate_flat_dem(self.minCoord,
-                                               self.maxCoord,
-                                               self.resolution,
-                                               self.surfElevation,
-                                               self.scaleDIM)
-
-            np.savetxt(self._demfile, self.dem)
-            # Build Mesh
-            self.badlands_model.build_mesh(self._demfile, verbose=False)
-
-            self.badlands_model.input.disp3d = True  # enable 3D displacements
-            self.badlands_model.input.region = 0  # TODO: check what this does
-            self.badlands_model.input.tStart = self.time_years
-            self.badlands_model.tNow = self.time_years
-
-            # Override the checkpoint/display interval in the Badlands model to
[…]
