+++ UWGeodynamics/surfaceProcesses.py
@@ -1,10 +1,18 @@
 from __future__ import print_function,  absolute_import
 import abc
+import underworld as uw
 import underworld.function as fn
+import numpy as np
+import sys
+from scipy.ndimage.filters import gaussian_filter
 from .scaling import nonDimensionalize as nd
+from .scaling import Dimensionalize
+from .scaling import UnitRegistry as u
+from tempfile import gettempdir
 
 
 ABC = abc.ABCMeta('ABC', (object,), {})
+_tempdir = gettempdir()
 
 
 class SurfaceProcesses(ABC):
@@ -33,56 +41,342 @@ class SurfaceProcesses(ABC):
 
 
 class Badlands(SurfaceProcesses):
-    """ A wrapper class for Badlands Linkage"""
+    """ A wrapper class for Badlands """
 
     def __init__(self, airIndex,
                  sedimentIndex, XML, resolution, checkpoint_interval,
-                 surfElevation=0., verbose=True, Model=None, restartFolder=None,
-                 restartStep=None, timeField=None, surfaceTracers=None):
+                 surfElevation=0., verbose=True, Model=None,
+                 restartFolder=None, restartStep=None, timeField=None,
+                 minCoord=None, maxCoord=None):
         try:
-
             import pyBadlands
 
-        except ImportError :
+        except ImportError:
             raise ImportError("""pyBadlands import as failed. Please check your
                               installation, PYTHONPATH and PATH environment
                               variables""")
 
+        self.verbose = verbose
+        self.restartStep = restartStep
+        self.restartFolder = restartFolder
+
+        self.Model = Model
         self.airIndex = airIndex
         self.sedimentIndex = sedimentIndex
-        self.XML = XML
-        self.resolution = resolution
-        self.checkpoint_interval = checkpoint_interval
-        self.surfElevation = surfElevation
-        self.verbose = verbose
-        self.restartFolder = restartFolder
-        self.restartStep = restartStep
-        self.surfaceTracers = surfaceTracers
-
+        self.resolution = nd(resolution)
+        self.surfElevation = fn.Function.convert(nd(surfElevation))
+        self.checkpoint_interval = nd(checkpoint_interval)
         self.timeField = timeField
-        self.Model = Model
+        self.XML = XML
+        self.time_years = 0.
+
+        if minCoord:
+            self.minCoord = tuple([nd(val) for val in minCoord])
+        else:
+            self.minCoord = self.Model.mesh.minCoord
+
+        if maxCoord:
+            self.maxCoord = tuple([nd(val) for val in maxCoord])
+        else:
+            self.maxCoord = self.Model.mesh.maxCoord
+
+        if self.mesh.dim == 2:
+            self.minCoord = (self.minCoord[0], self.minCoord[0])
+            self.maxCoord = (self.maxCoord[0], self.maxCoord[0])
 
     def _init_model(self):
 
-        from .linkage import SPM
+        if uw.mpi.rank == 0:
+            from pyBadlands.model import Model as BadlandsModel
+            self.badlands_model = BadlandsModel()
+            self.badlands_model.load_xml(self.XML)
 
-        self.mesh = self._Model.mesh
-        self.velocityField = self._Model.velocityField
-        self.swarm = self._Model.swarm
-        self.materialField = self._Model.materialField
+            if self.restartStep:
+                self.badlands_model.input.restart = True
+                self.badlands_model.input.rstep = self.restartStep
+                self.badlands_model.input.rfolder = self.restartFolder
+                self.badlands_model.input.outDir = self.restartFolder
+                self.badlands_model.outputStep = self.restartStep
 
[…]
