### 2019-03-06 ba5c0e2
Fix series of issues
