+++ UWGeodynamics/surfaceProcesses.py
@@ -5,6 +5,7 @@ import underworld.function as fn
 import numpy as np
 import sys
 from scipy.ndimage.filters import gaussian_filter
+from scipy.interpolate import griddata, interp1d
 from .scaling import nonDimensionalize as nd
 from .scaling import Dimensionalize
 from .scaling import UnitRegistry as u
@@ -60,7 +61,6 @@ class Badlands(SurfaceProcesses):
         self.restartStep = restartStep
         self.restartFolder = restartFolder
 
-        self.Model = Model
         self.airIndex = airIndex
         self.sedimentIndex = sedimentIndex
         self.resolution = nd(resolution)
@@ -69,22 +69,25 @@ class Badlands(SurfaceProcesses):
         self.timeField = timeField
         self.XML = XML
         self.time_years = 0.
+        self.minCoord = minCoord
+        self.maxCoord = maxCoord
+        self.Model = Model
 
-        if minCoord:
-            self.minCoord = tuple([nd(val) for val in minCoord])
+    def _init_model(self):
+
+        if self.minCoord:
+            self.minCoord = tuple([nd(val) for val in self.minCoord])
         else:
             self.minCoord = self.Model.mesh.minCoord
 
-        if maxCoord:
-            self.maxCoord = tuple([nd(val) for val in maxCoord])
+        if self.maxCoord:
+            self.maxCoord = tuple([nd(val) for val in self.maxCoord])
         else:
             self.maxCoord = self.Model.mesh.maxCoord
 
-        if self.mesh.dim == 2:
+        if self.Model.mesh.dim == 2:
             self.minCoord = (self.minCoord[0], self.minCoord[0])
             self.maxCoord = (self.maxCoord[0], self.maxCoord[0])
-
-    def _init_model(self):
 
         if uw.mpi.rank == 0:
             from pyBadlands.model import Model as BadlandsModel
@@ -122,7 +125,8 @@ class Badlands(SurfaceProcesses):
 
             # Override the checkpoint/display interval in the Badlands model to
             # ensure BL and UW are synced
-            self.badlands_model.input.tDisplay = self.checkpoint_interval
+            self.badlands_model.input.tDisplay = (
+            Dimensionalize(self.checkpoint_interval, u.years).magnitude)
 
             # Set Badlands minimal distance between nodes before regridding
             self.badlands_model.force.merge3d = (
@@ -135,13 +139,13 @@ class Badlands(SurfaceProcesses):
             # It would be nice if this wasn't the case.
             self.badlands_model.force.next_display = 0
 
-        uw.mpi.Barrier()
+        uw.mpi.barrier()
 
         self._disp_inserted = False
 
         # Transfer the initial DEM state to Underworld
         self._update_material_types()
-        uw.mpi.Barrier()
+        uw.mpi.barrier()
 
     def _generate_dem(self):
         """
@@ -182,7 +186,7 @@ class Badlands(SurfaceProcesses):
         dem[:, 0] = Dimensionalize(coordsX.flatten(), u.meter).magnitude
         dem[:, 1] = Dimensionalize(coordsY.flatten(), u.meter).magnitude
 
-        coordsZ = self.elevation.evaluate(dem[:, :1])
+        coordsZ = self.surfElevation.evaluate(dem[:, :1])
 
         dem[:, 2] = Dimensionalize(coordsZ.flatten(), u.meter).magnitude
         return dem
@@ -197,24 +201,26 @@ class Badlands(SurfaceProcesses):
         np_surface = None
         if uw.mpi.rank == 0:
             rg = self.badlands_model.recGrid
-            if self.mesh.dim == 2:
+            if self.Model.mesh.dim == 2:
                 zVals = rg.regZ.mean(axis=1)
                 np_surface = np.column_stack((rg.regX, zVals))
 
[…]
