### 2019-03-08 5dc45ca
Alias uwmpi to uw.mpi for UW version >= 2.8
