+++ UWGeodynamics/Underworld_extended/_meshvariable.py
@@ -8,6 +8,7 @@ from UWGeodynamics.scaling import Dimensionalize
 from UWGeodynamics.scaling import nonDimensionalize
 from UWGeodynamics.scaling import UnitRegistry as u
 from UWGeodynamics.version import git_revision as __git_revision__
+from UWGeodynamics import uwmpi
 
 
 class MeshVariable(uw.mesh.MeshVariable):
@@ -206,7 +207,7 @@ class MeshVariable(uw.mesh.MeshVariable):
         True
 
         >>> # clean up:
-        >>> if uw.mpi.rank == 0:
+        >>> if uwmpi.rank == 0:
         ...     import os;
         ...     os.remove( "saved_mesh_variable.h5" )
 
