### 2019-03-11 dfc99d0
Rely on mpi4y communicator directly...
