### 2019-03-11 12e41de
Fix particle state scaling
