### 2019-03-13 45c7019
Fix CompositeViscosity, the parent class was not initialised which caused all sorts of errors, one of them being that the firstIter attributs was not present. This should solve issue #98 raised by @mitchellmcm27
