### 2019-03-13 b1acd86
remove write to shapefile
