### 2019-03-14 75d3d1b
Refactor pre and post solve hooks, now use an OrderedDict which can be extended to store callback functions, more documentation to come...
