### 2019-03-14 99d8014
Solver can now be rebuilt using the Model._rebuild_solver flag, custom options set by the users are conserved
