+++ UWGeodynamics/_model.py
@@ -739,8 +739,17 @@ class Model(Material):
     @property
     def solver(self):
         # Get a solver
+        # In some cases, it is necessary to rebuild the solver,
+        # This can be done by switching the _rebuild_solver flag
+        # to true.
         if not self._solver or self._rebuild_solver:
+            if self._rebuild_solver and self._solver:
+                # Save current options in a dictionary
+                options = _solver_options_dictionary(self._solver)
             self._solver = uw.systems.Solver(self.stokes_SLE)
+            if self._rebuild_solver and self._solver:
+                # Apply saved options on *new* solver
+                _apply_saved_options_on_solver(self._solver, options)
         return self._solver
 
     @property
@@ -2731,3 +2740,26 @@ def _get_output_units(*args):
             return arg.units
 
     return rcParams["time.SIunits"]
+
+
+def _solver_options_dictionary(solver):
+    """Return a dictionary of all the solver options"""
+    dd = {}
+    for key, val in solver.options.__dict__.items():
+        dd2 = {}
+        for key2, val2, in val.__dict__.items():
+            dd2[key2] = val2
+        dd[key] = dd2
+    return dd
+
+
+def _apply_saved_options_on_solver(solver, options):
+    """Apply options on a solver
+
+    solver: solver
+    options: python dictionary
+    """
+    for key, val in options.items():
+        for key2, val2 in val.items():
+            solver.options.__dict__[key].__dict__[key2] = val2
+    return solver
