### 2019-03-14 4fd259a
Fix Moving Wall, tested in parallel using 4Cpus...more to come
