+++ UWGeodynamics/_boundary_conditions.py
@@ -369,14 +369,14 @@ class VelocityBCs(BoundaryConditions):
 
         if isinstance(condition, MovingWall):
             condition.wall = nodes
-            set_, axis = condition.move_wall()
+            set_, axis = condition.get_wall_indices()
             func = condition.velocityFn
 
             # Intersect of wall nodes and current local domain
             ISet = uw.mesh.FeMesh_IndexSet(
                 self.Model.mesh, topologicalIndex=0,
                 size=self.Model.mesh.nodesGlobal,
-                fromObject=set_.data)
+                fromObject=set_)
 
             for dim in range(self.Model.mesh.dim):
                 if ISet.data.size > 0:
