### 2019-03-14 0529167
LithoPress now return the lithostatic pressure field as a MeshVariable Add initialize_pressure_to lithostatic method Add lithostatic_pressureField property That should allow passing the lithostatic_pressureField as an intial stressBCs, note that the condition is not updated (It remains as the initial lithostaticField)
