### 2019-03-14 738cd29
Launch jupyter lab, get python3 badlands
