+++ docker-dev/Dockerfile
@@ -21,7 +21,7 @@ RUN pip3 install --no-cache-dir pandas \
                 git+https://github.com/awickert/gFlex.git
 
 # pyBadlands serial
-RUN git clone https://github.com/rbeucher/pyBadlands_serial.git
+RUN git clone -b python3 https://github.com/rbeucher/pyBadlands_serial.git
 RUN cd /opt/pyBadlands_serial/pyBadlands/libUtils && make
 RUN pip3 install -e /opt/pyBadlands_serial
 
@@ -63,5 +63,5 @@ ENV PYTHONPATH /workspace/user_data/Underworld:/workspace/user_data/UWGeodynamic
 WORKDIR /workspace
 
 # launch notebook
-CMD ["jupyter", "notebook", "--ip='0.0.0.0'", "--no-browser"]
+CMD ["jupyter", "lab", "--ip='0.0.0.0'", "--no-browser"]
 
