### 2019-03-14 f2a2e1e
Add nProcs as an alias for uw.mpi.size
