### 2019-03-14 7e12b02
Fix issues pointed out by @clairemallard in issue #96
