### 2019-03-14 fcb1aa5
Fix error in parallel
