### 2019-03-14 ccc8d70
Fix issue #99, thanks @plovely
