### 2019-03-14 4ec998d
Add legacy contour to docker, get latest matplotlib
