### 2019-03-14 583c65c
Some work on remeshing, dynamic remeshing does not work though... more coming
