+++ UWGeodynamics/_remeshing.py
@@ -1,46 +1,97 @@
 import numpy as np
 from .scaling import nonDimensionalize as nd
+from .Underworld_extended import MeshVariable
+from scipy.interpolate import interp1d
+import copy
 
 
-def remesh(mesh, x=None, y=None, z=None):
-    nx, ny = mesh.elementRes
+class ReMesher(object):
 
-    def new_points(intervals, elements):
-        pts = []
-        for idx in range(len(intervals) - 1):
-            pts.append(np.linspace(intervals[idx], intervals[idx + 1], elements[idx] + 1))
-        pts = np.unique(np.hstack(pts))
-        pts.sort()
-        return pts
+    def __init__(self, Model, x, y, z, reset=False, adaptive=False):
 
-    def check_vals(intervals, elements, axis):
-        if ((intervals[0] != mesh.minCoord[axis]) or
-            (intervals[-1] != mesh.maxCoord[axis])):
-            raise ValueError("""Intervals do not match mesh extent""")
+        self.mesh = Model.mesh
+        self.initial_mesh = copy.copy(self.mesh)
+        self.x = x
+        self.y = y
+        self.z = z
+        self.reset = reset
 
-        if np.sum(np.array(elements)) != mesh.elementRes[axis]:
-            raise ValueError("""Total nb of elements do not match the nb of elements in the mesh""")
+        if adaptive:
+            Model.post_solve_functions["Remesher"] = self.remesh
 
-    if x:
-        intervals, elements = x
-        intervals = [nd(val) for val in intervals]
-        elements = [nd(val) for val in elements]
-        check_vals(intervals, elements, 0)
-        pts = new_points(intervals, elements)
+    def reset_mesh(self):
+        with self.mesh.deform_mesh():
+            self.mesh.data[...] = self.initial_mesh.data[...]
 
-        with mesh.deform_mesh():
-            new_vals = np.tile(pts, ny + 1)
-            mesh.data[:, 0] = new_vals[mesh.data_nodegId.flatten()]
+    def remesh(self):
 
-    if y:
-        intervals, elements = y
-        intervals = [nd(val) for val in intervals]
-        elements = [nd(val) for val in elements]
-        check_vals(intervals, elements, 1)
-        pts = new_points(intervals, elements)
+        if self.reset:
+            self.reset_mesh()
 
-        with mesh.deform_mesh():
-            vals = np.repeat(pts, nx + 1)
-            mesh.data[:, 1] = vals[mesh.data_nodegId.flatten()]
-    return mesh
+        nx, ny = self.mesh.elementRes
+
+        def new_points(intervals, elements):
+            pts = []
+            for idx in range(len(intervals) - 1):
+                pts.append(np.linspace(intervals[idx], intervals[idx + 1], elements[idx] + 1))
+            pts = np.unique(np.hstack(pts))
+            pts.sort()
+            return pts
+
+        def check_vals(intervals, elements, axis):
+            if ((intervals[0] != self.mesh.minCoord[axis]) or
+                (intervals[-1] != self.mesh.maxCoord[axis])):
+                raise ValueError("""Intervals do not match mesh extent""")
+
+            if np.sum(np.array(elements)) != self.mesh.elementRes[axis]:
+                raise ValueError("""Total nb of elements do not match the nb of elements in the mesh""")
+
+        def new_points_from_field(field, axis):
+            nx, ny = self.mesh.elementRes
+            A = field.data.reshape((nx + 1, ny + 1))
+            B = A.max(axis=axis)
+            C = B / B.max()
+            D = C.cumsum()
+            E = D - D[0]
+            E = E / E.max()
+            if axis == 1:
+                coords = self.mesh.data[:, axis].reshape((nx + 1, ny + 1))[:, -1]
+            if axis == 0:
[…]
