### 2019-03-15 f445650
Now that the scaling module has been integrated into UW, rely on it directly. Fall back to local module if the UW version does not have the scaling (for back-compatibility) nonDimensionalize -> non_dimensionalise (Old name shows a warning Dimensionalize -> dimensionalise (Old name shows a warning
