### 2019-03-15 5d97f2b
Fix mistake in _remeshing, probably a copy and paste
