### 2019-03-15 1e73da2
Adapt time units in log outputs
