### 2019-03-15 6d4fce3
Fix pb when rebuilding new solver
