+++ UWGeodynamics/_model.py
@@ -762,6 +762,7 @@ class Model(Material):
             self._solver = uw.systems.Solver(self.stokes_SLE)
             if self._rebuild_solver and self._solver:
                 # Apply saved options on *new* solver
+                options = _solver_options_dictionary(self._solver)
                 _apply_saved_options_on_solver(self._solver, options)
         return self._solver
 
