### 2019-03-17 f3fed9d
Fix material as boundary condition
