### 2019-03-18 16753a2
Fix material as bc in parallel
