+++ UWGeodynamics/_model.py
@@ -1465,9 +1465,8 @@ class Model(Material):
         mat = self.projMaterialField.evaluate(self.mesh)
         # Get nodes corresponding to material
         mask = (mat == Decimal(material.index))
-        if mask:
-            nodes = np.arange(0, self.mesh.nodesDomain)[mat == mask]
-            return uw.mesh.FeMesh_IndexSet(self.mesh, topologicalIndex=0,
+        nodes = np.arange(0, self.mesh.nodesDomain)[mask.flatten()]
+        return uw.mesh.FeMesh_IndexSet(self.mesh, topologicalIndex=0,
                                            size=self.mesh.nodesGlobal,
                                            fromObject=nodes)
 
