### 2019-03-18 e244738
Fix pb when evaluating at the boundary in parallel. The lithostatic pressure function now return a field variable that lives on the mesh, not on the submesh...Pressure is initialized by evaluating the field on the subMesh
