### 2019-03-18 ac016ab
Fix lithostatic BCs issue in parallel
