### 2019-03-19 c4d9a4e
Apply patch to fix #100
