### 2019-03-20 b6ccdcc
Add remeshing in 3D, there is still some pb with the y axis though...
