+++ UWGeodynamics/lithopress/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import print_function,  absolute_import
-from .lithopress import LithostaticPressure
+from .lithopress import Lithostatic_pressure
 __version__ = "0.1"
