### 2019-03-21 8bf5759
Get rid of the Shape class, make all shapes inherit from fn.Function
