+++ UWGeodynamics/shapes.py
@@ -5,37 +5,7 @@ import underworld.function as fn
 from UWGeodynamics import non_dimensionalise as nd
 
 
-class Shape(object):
-    """Base Class for Shape objects"""
-
-    def __init__(self):
-        self._fn = None
-        return
-
-    @property
-    def fn(self):
-        return self._fn
-
-    @fn.setter
-    def fn(self, value):
-        self._fn = value
-
-    def __and__(self, B):
-        newShape = Shape()
-        newShape.fn = self._fn & B._fn
-        return newShape
-
-    def __add__(self, B):
-        newShape = CombinedShape([self, B])
-        return newShape
-
-    def __or__(self, B):
-        newShape = Shape()
-        newShape.fn = self._fn | B._fn
-        return newShape
-
-
-class Polygon(Shape):
+class Polygon(fn.Function):
     """Polygon Shape Class"""
 
     def __init__(self, vertices):
@@ -56,9 +26,11 @@ class Polygon(Shape):
         self.bottom = max([y for x, y in self.vertices])
         vertices = [(nd(x), nd(y)) for x, y in self.vertices]
         self._fn = uw.function.shape.Polygon(np.array(vertices))
+        super(Polygon, self).__init__(argument_fns=None)
+        self._fncself = self._fn._fncself
 
 
-class HalfSpace(Shape):
+class HalfSpace(fn.Function):
     """ Class to define a HalfSpace
 
         A plan defined by a normal vector is used to split the space
@@ -99,8 +71,6 @@ class HalfSpace(Shape):
 
         self.reverse = reverse
 
-    @property
-    def _fn(self):
         coords = fn.input()
         new_coords = coords - self.origin
         func = fn.math.dot(self.normal, new_coords)
@@ -111,84 +81,12 @@ class HalfSpace(Shape):
         else:
             conditions = [(func >= 0., True), (func < 0., False)]
 
-        return fn.branching.conditional(conditions)
+        self._fn = fn.branching.conditional(conditions)
+        super(HalfSpace, self).__init__(argument_fns=None)
+        self._fncself = self._fn._fncself
 
 
-class CombinedShape(Shape):
-    """CombinedShape class"""
-
-    def __init__(self, shapes):
-        """Combine a list of shapes into one
-
-        Parameters
-        ----------
-
-        shapes : list of Shape objects
-
-        Returns
-        -------
-
-        An UWGeodynamics Shape object
-        """
-        self.shapes = shapes
-        self.top = max([shape.top for shape in shapes])
-        self.bottom = min([shape.top for shape in shapes])
-
-    @property
-    def _fn(self):
-        import operator
-        import functools
-        self._fnlist = []
-        for shape in self.shapes:
-            self._fnlist.append(shape._fn)
-        func = functools.reduce(
-            operator.or_,
-            self._fnlist,
-            fn.misc.constant(False))
-        return func
-
-
-class MultiShape(CombinedShape):
-
-    def __init__(self, shapes):
-
-        Warning("MultiShape is now deprecated, use CombinedShape instead")
-        super(MultiShape, self).__init__(shapes)
-
-
-class IntersectShape(Shape):
-    """IntersectShape class"""
-
-    def __init__(self, shapes):
-        """ Take the intersection of some shapes
-
-        Parameters
-        ----------
-
-        shapes : A list of Shapes
-
-        Returns
-        -------
-
-        An UWGeodynamics Shape object
-        """
-        self.shapes = shapes
-
-    @property
-    def _fn(self):
-        import operator
-        import functools
[…]
