### 2019-03-22 e00bbff
Passive Tracers now inherits from Swarm
This essentially change nothing except when plotting using glucifer,
user can now directly pass the tracers as a swarm, not need to get tracer.swarm...
