### 2019-03-25 b3b1e47
Get Rid of Coordinates in Passive Tracers
