+++ UWGeodynamics/_utils.py
@@ -213,11 +213,6 @@ class PassiveTracers(Swarm):
                 globalCount = len(dset)
                 dim = self.mesh.dim
 
-            string += "\t<Attribute Type=\"Vector\" Center=\"Node\" Name=\"Coordinates\">\n"
-            string += """\t\t\t<DataItem Format=\"HDF\" NumberType=\" Float\"
-                         Precision=\"8\" Dimensions=\"{0} {1}\">{2}:/data</DataItem>\n""".format(globalCount, dim, swarm_fname)
-            string += "\t</Attribute>\n"
-
             # Write the footer to the xmf
             string += uw.utils._xdmffooter()
 
