### 2019-03-26 69ea258
The swarmvarshema do not check the swarmvariable type. Change that using a custom version of the function. Need to be merged into UW once tested.
