### 2019-03-26 79cbaa2
Run the tests and fix some issues in the examples
