### 2019-03-27 c258150
Fix issue #101, Fix dataType in XDMF files
