### 2019-03-27 #1136 @flohorovicic
The provided software and description provides a really low entry point for the use of numerical simulations for typical applications in geodynamics. Especially the interface to Python makes using this software with all its underlying potential now very easy and I can see a wide potential for the use in teaching as well as research.

One aspect that is a bit puzzling is the separation from the underworld2 project (https://github.com/underworldcode/underworld2). When going through the examples and looking online, I commonly ended up on information referring to underworld2 - especially as it also has a Python interface and some (also very interesting) example notebooks. I ended up running several of these notebooks before realising it is a different repository... It is not directly obvious why they are two different projects.

Suggestion: extend information on the UWGeodynamics readme file, clarify difference to underworld2 and include also installation information directly there (and a link to the notebooks on binder).

**Installation**

The suggested installation over Docker worked without any problems. Small suggestion: add some notes on how to stop a container (and even delete an image, as it can take a lot of space - or at least mention the required space). People not using Docker on a regular basis would benefit from it (even if this information is, of course, available with a simple search - it would lower the entry level of using Docker).

**Live Notebooks on binder**

This is really an excellent possibility to get started directly with the software without a local installation and most examples worked (apart from the Tutorials). Please also include information on the main readme page (apart from the launch binder link at the top - for people who are not familiar with binder and this possibility...).

**Documentation**

The online documentation on https://uwgeodynamics.readthedocs.io/en/latest/ is really helpful! Is it possible to also include a link to this documentation directly in the paper?

The tutorials and examples provide a very good starting point to run the software. The provided example notebooks contain simulations of typical applications one would want to model with the software. The description in these notebooks is short, but sufficient to understand the context of the examples (typically also with references to related work).

**Community guidelines**

_Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_

I did not find community guidelines and possibilities for contribution. However, questions on the Issue tracker are addressed very quickly, so support is obvious (but again, some more information on the readme page would help).

**Version**

_Version: Does the release version given match the GitHub release (v1.0.0)?_

The DOI version on Zenodo is UWGeodynamics v2.7.4

**Summary**

Most of my comments refer to simply more information in the readme file. Great project (or projects - with underworld2) overall, looking forward to using it more in the future in my class!
