### 2019-03-27 99257c5
If no restartStep nor restartFolder is specified for Badlands, use the Underworld timestep and the current badlands outputdir as restart Folder, This should facilitate chaining of pbs jobs
