### 2019-03-28 dcedb14
Add a Aspect ratio for 2d models
