+++ UWGeodynamics/surfaceProcesses.py
@@ -75,6 +75,7 @@ class Badlands(SurfaceProcesses):
         self.time_years = 0.
         self.minCoord = minCoord
         self.maxCoord = maxCoord
+        self.aspectRatio2d = aspectRatio2d
         self.Model = Model
 
     def _init_model(self):
@@ -90,8 +91,8 @@ class Badlands(SurfaceProcesses):
             self.maxCoord = self.Model.mesh.maxCoord
 
         if self.Model.mesh.dim == 2:
-            self.minCoord = (self.minCoord[0], aspectRatio2d*self.minCoord[0])
-            self.maxCoord = (self.maxCoord[0], aspectRatio2d*self.maxCoord[0])
+            self.minCoord = (self.minCoord[0], self.aspectRatio2d*self.minCoord[0])
+            self.maxCoord = (self.maxCoord[0], self.aspectRatio2d*self.maxCoord[0])
 
         if rank == 0:
             from pyBadlands.model import Model as BadlandsModel
