### 2019-03-28 482e7c8
Add some instructions on how to contribute
