### 2019-03-28 2cc9fdb
Aspect ratio2d (#104)
* Update badlands docker

underworld v2.6.0 is a bit of issues with singularity and mpi

* Add a Aspect ratio for 2d models

* Update surfaceProcesses.py
