### 2019-04-01 3196752
Make Some changes to the README and documentation to address JOSS review
