### 2019-04-01 eaa31df
Add some information to the README
Update Readme and documentation

Fix link

Change section on Underworld/UWGeodynamics in the Readme
