### 2019-04-01 underworldcode/UWGeodynamics#106 @rbeucher
I have added an Installation section to the README file. I have also added some information on how to create, list, delete docker images and containers...31967526b7399237d8b7639eae2633069a10b7bf
