### 2019-04-01 underworldcode/UWGeodynamics#107 @rbeucher
I have added a small paragraph about Binder together with an explicit link to the live Binder:
31967526b7399237d8b7639eae2633069a10b7bf

Quick Start / Testing
We provide a docker container via binder. This is a quick solution to get you started and run the examples and tutorials without installing anything on your machine. That is a good way to see if the software can actually be useful to you. The ressource are however limited and you should not try to run model with high resolution. 3D models can not be run in the binder.
