### 2019-04-01 underworldcode/UWGeodynamics#108 @rbeucher
Add link to the readTheDocs documentation in the JOSS Paper
