### 2019-04-01 underworldcode/UWGeodynamics#109 @rbeucher
I have added an How to Contribute section to the README file (Also on readTheDocs).
This includes some information on how to seek help, how to submit codes etc..
