### 2019-04-01 underworldcode/UWGeodynamics#110 @rbeucher
Add a section to explain the version number of the software:

Note on versioning
Since version 1.0 The Underworld development team has decided to match the UWGeodynamics version number with the latest supported version of Underworld. UWGeodynamics v2.7 is then supporing Underworld up to version 2.7.

The third number is used for UWGeodynamics only (v2.7.1, v2.7.2 etc.)

The development branch is based on the current Underworld development branch.

The Current release (DOI citable):
