### 2019-04-01 #1136 @rbeucher
Hi  @flohorovicic

Thank you very much for taking the time to review our project. Your comments and suggestions were very useful.

I have open a series of issues on the repository where I quickly list the changes made.
They essentially affect the documentation in the README and on READTHEDOCS.
All the changes were squashed into one single commit. https://github.com/underworldcode/UWGeodynamics/commit/31967526b7399237d8b7639eae2633069a10b7bf

The first point was about clarifying the difference between Underworld and UWGeodynamics.
I have added a short paragraph in the readme: https://github.com/underworldcode/UWGeodynamics/issues/105

**Installation**

I have copied the Installation section from ReadTheDocs to the README file
I have also added some instructions on how to deal with docker images and containers
https://github.com/underworldcode/UWGeodynamics/issues/106

**Documentation**
I have added a link to the documentation in the JOSS paper.
https://github.com/underworldcode/UWGeodynamics/issues/108

**Binder**
I have added a short paragraph and a link to the binder
https://github.com/underworldcode/UWGeodynamics/issues/107

**Community guidelines**

I have added some guidelines on how to contribute, raise issues and seek support in the README and
the readthedocs website.
https://github.com/underworldcode/UWGeodynamics/issues/109

**Version**
The version is currently 2.7.5. We changed the version number system to match the version of the latest Underworld API supported. I have added a paragraph in the README and ReadTheDocs.
I have not made significant changes since v1.0 beside fixing bugs etc.
https://github.com/underworldcode/UWGeodynamics/issues/110
@lheagy I am happy to draft a new released once the review process is finished.

Thanks again for your review and support!

Romain Beucher
