### 2019-04-03 #1136 @flohorovicic
Hi @rbeucher,

Great to hear that the comments were helpful!

The clarifications are very useful, especially on the distinction to underworld and the section on contributing. Looking forward to seeing this project evolve and being used in research and teaching eben more!

@lheagy: All fine from my side!

Kind regards,

     — Florian —
