### 2019-04-04 5b599fc
Temperature Field was not updated properly before being passed to the viscosity components. This fix #112
