### 2019-04-04 de59d93
Up version to 2.7.6
