### 2019-04-06 #1136 @lheagy
Many thanks @flohorovicic for taking the time to review!!

@rbeucher: One clarifying question:

- [ ] Do you have a contributor guide or similar?  I did not see a doc in the repository or page in the documentation that satisfies the criteria
   > Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
