### 2019-04-08 3ab884c
Fix for pip installations
add pandas and pint as dependencies
