+++ requirements.txt
@@ -2,3 +2,5 @@ numpy==1.15.1
 scipy==1.1.0
 mpi4py==3.0.0
 h5py==2.8.0
+pandas==0.24.0
+pint>=0.8.1
