### 2019-04-09 750192f
Fix typo in joss paper
