### 2019-04-10 #1136 @lheagy
@rbeucher: can you please generate a new release and archive your software on [zenodo](http://zenodo.org) or similar? Please make sure that the title and author list on the zenodo archive match the JOSS paper. Once you have done this, please post the version number and doi here and we can proceed with accepting your paper.
