### 2019-04-11 #1136 @rbeucher
Hi @lheagy

Here it is

Romain Beucher, Louis Moresi, Julian Giordani, John Mansour, Dan Sandiford, Rebecca Farrington, … Sara Morón. (2019, April 11). UWGeodynamics: A teaching and research tool for numerical geodynamic modelling (Version v2.7.7). Zenodo. http://doi.org/10.5281/zenodo.2636105
