### 2019-04-11 #1136 @rbeucher
I have merged it
