### 2019-04-11 #1136 @labarba
Congratulations @rbeucher, your JOSS paper is published!

Many thanks to our editor: @lheagy, and the reviewer: @flohorovicic — thanks also to @tth030 for the efforts! 🙏
